compartmentalise the learn link derivation
