+++ flucoma/doc/transformers.py
@@ -6,10 +6,11 @@
 # under the European Union’s Horizon 2020 research and innovation programme
 # (grant agreement No 725899).
 
+import copy
 import logging
+from flucoma.doc.learn import derive_learn_link
 from flucoma.doc import logger
 from collections import OrderedDict
-import copy
 from functools import reduce
 
 """
@@ -42,34 +43,10 @@ def tidy_split(string,separator=','):
     return map(lambda x: x.strip(),
         filter(lambda x: len(x), string.split(separator))
     )
-
-def derive_learn_link(object_name):
-    url_map = {
-        'bufampfeature' : 'ampfeature',
-        'bufnoveltyfeature' : 'noveltyfeature',
-        'bufonsetfeature' : 'onsetfeature',
-        'bufspectralshape' : 'spectralshape',
-        'bufchroma' : 'chroma',
-        'bufloudness' : 'loudness',
-        'bufmelbands' : 'melbands',
-        'bufmfcc' : 'mfcc',
-        'bufpitch' : 'pitch',
-        'bufhpss' : 'hpss',
-        'bufsines' : 'sines',
-        'buftransients' : 'transients',
-        'bufampgate' : 'ampgate',
-        'bufampslice' : 'ampslice',
-        'bufnoveltyslice' : 'noveltyslice',
-        'bufonsetslice' : 'onsetslice',
-        'buftransientslice' : 'transientslice',
-        'bufaudiotransport' : 'audiotransport'
-    }
-    return url_map.get(object_name.lower()) or object_name.lower()
  
 def default_transform(object_name, data):
-    
     data['client_name'] = object_name 
-    data['learn_url'] = derive_learn_link(object_name)
+    data['learn_url'] = 'https://learn.flucoma.org/reference/' + derive_learn_link(object_name)
     data['category'] = []
     data['keywords'] = []
     data['module'] = 'fluid decomposition'
